Refactor project structure and update dependencies
- Update Expo SDK dependencies to latest versions for improved compatibility.
- Add expo-font plugin to app.json for font management.
- Introduce bun.lock for dependency management, replacing package-lock.json.
- Refactor GA imports to use a new libs directory.
- Adjust styles in BlogList component for better layout and responsiveness.
- Update TypeScript configuration and VSCode settings for improved development experience.
- Minor adjustments to animated components for performance enhancements.
