+++ .vscode/settings.json
@@ -2,5 +2,6 @@
   "yaml.schemas": {
     "https://www.graphql-code-generator.com/config.schema.json": ["codegen.yml"],
     "https://json.schemastore.org/github-workflow.json": "file:///Users/sugiyamahironori/react-native/expo/me/.github/workflows/deploy.yml"
-  }
+  },
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
