+++ package.json
@@ -19,29 +19,29 @@
     "@react-navigation/native": "^7.0.0",
     "dayjs": "^1.11.9",
     "expo": "~54.0.0",
-    "expo-constants": "~17.0.0",
-    "expo-font": "~13.3.0",
-    "expo-image": "~2.3.0",
-    "expo-linking": "~7.1.0",
+    "expo-constants": "~18.0.13",
+    "expo-font": "~14.0.11",
+    "expo-image": "~3.0.11",
+    "expo-linking": "~8.0.11",
     "expo-router": "~6.0.0",
-    "expo-status-bar": "~2.2.0",
+    "expo-status-bar": "~3.0.9",
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "react-native": "0.81.4",
-    "react-native-gesture-handler": "~2.25.0",
-    "react-native-reanimated": "~4.0.0",
-    "react-native-safe-area-context": "~5.4.0",
-    "react-native-screens": "~4.11.0",
-    "react-native-svg": "~15.11.0",
-    "react-native-web": "~0.20.0",
-    "react-native-worklets": "~0.7.0",
+    "react-native": "0.81.5",
+    "react-native-gesture-handler": "~2.28.0",
+    "react-native-reanimated": "~4.1.1",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0",
+    "react-native-svg": "15.12.1",
+    "react-native-web": "^0.21.0",
+    "react-native-worklets": "0.5.1",
     "react-use": "^17.4.2",
-    "typescript": "~5.8.0"
+    "typescript": "~5.9.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "@types/react": "~19.0.0",
     "@types/gtag.js": "^0.0.18",
+    "@types/react": "~19.1.10",
     "@typescript-eslint/eslint-plugin": "^6.6.0",
     "@typescript-eslint/parser": "^6.6.0",
     "ajv": "^8.12.0",
@@ -63,9 +63,6 @@
   },
   "private": true,
   "engines": {
-    "node": ">=20"
-  },
-  "peerDependencies": {
-    "typescript": "^5.0.0"
+    "node": ">=22"
   }
 }
