+++ src/app/_layout.tsx
@@ -7,7 +7,7 @@ import { ThemeProvider } from '@react-navigation/native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { CustomThemeProvider, useCustomTheme } from '@/components/context/CustomThemeProvider';
 import { useEffect } from 'react';
-import { pageview } from '@/lib/ga';
+import { pageview } from '@/libs/ga';
 import { useFonts, Inter_400Regular, Inter_700Bold } from '@expo-google-fonts/inter';
 
 export const unstable_settings = {
