+++ src/components/BlogList.module.css
@@ -1,7 +1,13 @@
-.list {
+/* .list {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(min(768px, 100%), 1fr));
   gap: 32px;
+  margin: 32px 0;
+} */
+
+.blog_list {
+  flex: 1;
+  align-items: center;
   margin: 32px 0;
 }
 
@@ -16,6 +22,44 @@
   margin: 8px;
   animation: fade-in 1s linear;
   transition: all 0.3s;
+}
+
+.card_image_container {
+  flex: 1;
+  max-width: 160px;
+  margin: 2px;
+  margin-left: 4px;
+  margin-right: 4px;
+}
+
+.card_image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.card_detail {
+  flex: 1;
+  overflow-x: hidden;
+  overflow-y: auto;
+}
+
+.card_title {
+  font-size: 18px;
+  font-weight: bold;
+  text-decoration: underline;
+}
+
+.card_content {
+  flex: 1;
+  padding: 4px;
+  justify-content: center;
+  align-items: center;
+}
+
+.card_date {
+  align-self: flex-end;
+  padding: 0 8px;
 }
 
 .card:hover {
