+++ src/components/BlogList.tsx
@@ -1,10 +1,11 @@
-import { StyleSheet, View } from 'react-native';
+import { View } from 'react-native';
 import { Image } from 'expo-image';
 
 import { NativeText } from '@/components/Text';
 import { ExternalLink } from '@/components/ExternalLink';
 import Rss from '@/constants/rss/rss.json';
 
+// @ts-expect-error
 import { unstable_styles } from './BlogList.module.css';
 
 type RssItemProp = (typeof Rss)[0];
@@ -13,13 +14,13 @@ export const BlogItem = ({ item }: { item: RssItemProp; index: number }) => {
   return (
     <ExternalLink key={item.id} href={item.link} asChild>
       <View style={unstable_styles.card}>
-        <View style={styles.cardImage}>
-          <Image style={styles.cardImage} source={{ uri: item.og }} contentFit='contain' />
+        <View style={unstable_styles.card_image_container}>
+          <Image style={unstable_styles.card_image} source={{ uri: item.og }} contentFit='contain' />
         </View>
-        <View style={styles.cardDetail}>
-          <NativeText style={styles.cardTitle}>{item.title}</NativeText>
-          <NativeText style={styles.cardContent}>{item.contentSnippet.substring(0, 100)}...</NativeText>
-          <NativeText style={styles.cardDate}>{item.isoData}</NativeText>
+        <View style={unstable_styles.card_detail}>
+          <NativeText style={unstable_styles.card_title}>{item.title}</NativeText>
+          <NativeText style={unstable_styles.card_content}>{item.contentSnippet.substring(0, 100)}...</NativeText>
+          <NativeText style={unstable_styles.card_date}>{item.isoData}</NativeText>
         </View>
       </View>
     </ExternalLink>
@@ -28,59 +29,10 @@ export const BlogItem = ({ item }: { item: RssItemProp; index: number }) => {
 
 export const BlogList = () => {
   return (
-    <View style={styles.list}>
+    <View style={unstable_styles.blog_list}>
       {Rss.map((item, index) => (
         <BlogItem key={item.id} {...{ item, index }} />
       ))}
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  list: {
-    flex: 1,
-    alignItems: 'center',
-    marginVertical: 32,
-  },
-  card: {
-    flexDirection: 'row',
-    width: '100%',
-    height: 160,
-    maxWidth: 768,
-    backgroundColor: '#000',
-    padding: 8,
-    paddingVertical: 12,
-    paddingBottom: 16,
-    borderRadius: 8,
-    margin: 8,
-    shadowColor: '#ddd',
-    shadowOffset: { width: 2, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 10,
-  },
-  cardTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    textDecorationLine: 'underline',
-  },
-  cardContent: {
-    flex: 1,
-    padding: 4,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  cardDate: {
-    alignSelf: 'flex-end',
-    paddingHorizontal: 8,
-  },
-  cardImage: {
-    flex: 1,
-    maxWidth: 160,
-    margin: 2,
-    marginRight: 4,
-  },
-  cardDetail: {
-    flex: 1,
-    overflow: 'hidden',
-  },
-});
